% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{roadmap}
\alias{roadmap}
\title{Default Variable-Keyword Roadmap}
\format{
A tibble with 3 rows and 2 variables:
\describe{
\item{variable}{Clinical measurement variable name}
\item{if_missing_search_for}{Diagnostic keyword to search in notes when structured data is missing}
}
}
\usage{
roadmap
}
\description{
Default mapping of clinical variables to diagnostic keywords for missing data search.
}
\examples{
data(roadmap)
roadmap

}
\keyword{datasets}
